Remove uploader from ReportConfig & update setters
