+++ src/main/java/ru/yandex/qatools/allure/jenkins/dsl/AllureReportPublisherContext.java
@@ -1,9 +1,13 @@
 package ru.yandex.qatools.allure.jenkins.dsl;
 
 import javaposse.jobdsl.dsl.Context;
+import ru.yandex.qatools.allure.jenkins.AllureReportDefaultUploader;
+import ru.yandex.qatools.allure.jenkins.AllureReportUploader;
 import ru.yandex.qatools.allure.jenkins.config.AllureReportConfig;
 import ru.yandex.qatools.allure.jenkins.config.PropertyConfig;
 import ru.yandex.qatools.allure.jenkins.config.ReportBuildPolicy;
+
+import java.util.List;
 
 /**
  * @author Marat Mavlutov <mavlyutov@yandex-team.ru>
@@ -13,9 +17,11 @@ class AllureReportPublisherContext implements Context {
     public static final String FAILURE_POLICY = "FAILURE";
 
     private AllureReportConfig config;
+    private AllureReportUploader uploader;
 
-    public AllureReportPublisherContext(AllureReportConfig config) {
+    public AllureReportPublisherContext(AllureReportConfig config, AllureReportUploader uploader) {
         this.config = config;
+        this.uploader = uploader;
     }
 
     public AllureReportConfig getConfig() {
@@ -43,8 +49,26 @@ class AllureReportPublisherContext implements Context {
         getConfig().setIncludeProperties(includeProperties);
     }
 
-    public void uploader(String uploader) {
-        this.getConfig().setUploader(uploader);
+    public AllureReportUploader getUploader() {
+        return uploader == null ? new AllureReportDefaultUploader() : uploader;
+    }
+
+    public void uploader(String uploaderName) {
+        AllureReportUploader selectedUploader = null;
+        List<AllureReportUploader> availableUploaders = AllureReportUploader.all();
+        for (AllureReportUploader availableUploader: availableUploaders) {
+            if (availableUploader.getDescriptor().getDisplayName().equals(uploaderName)) {
+                selectedUploader = availableUploader;
+                break;
+            }
+        }
+        if (selectedUploader == null) {
+            //throw new NullPointerException(String.format("There are no uploaders with name `%s`", uploaderName));
+        }
+        else {
+            this.uploader = selectedUploader;
+        }
+
     }
 
 }
