+++ src/test/java/ru/yandex/qatools/allure/jenkins/DslIntegrationTest.java
@@ -49,7 +49,7 @@ public class DslIntegrationTest {
         assertThat(allureReportPublisher.getConfig().getProperties(), hasSize(1));
         assertThat(allureReportPublisher.getConfig().getProperties().get(0).getKey(), equalTo("key"));
         assertThat(allureReportPublisher.getConfig().getProperties().get(0).getValue(), equalTo("value"));
-        assertThat(allureReportPublisher.getConfig().getUploader().getDescriptor().getDisplayName(), equalTo("Jenkins Master"));
+        assertThat(allureReportPublisher.getUploader().getDescriptor().getDisplayName(), equalTo("Jenkins Master"));
 
         assertThat(allureReportPublisher.getConfig().getReportBuildPolicy(), equalTo(ReportBuildPolicy.UNSTABLE));
         assertThat(allureReportPublisher.getConfig().getIncludeProperties(), equalTo(Boolean.TRUE));
